Why use constructors in service files?
This PR is opened in the form of a question. Why are constructors used
at all in the singleton module services? It's awkward that something is
exported initialized. If the constructor were exported, I could see it
being useful for testing and reusability, but right now it's just an
object serving as a layer of indirection.